bridge_adaptivity/requirements_base.txt
celery==4.1.0

Django==2.0.5
django-bootstrap3==9.0.0
django-fontawesome==0.3.1
django-cors-middleware==1.3.1
django-ordered-model==1.4.3
djangorestframework==3.7.7

edx-rest-api-client==1.7.1
lti==0.9.2
requests==2.11.1
psycopg2==2.7.3
shortuuid==0.5.0  # https://github.com/skorokithakis/shortuuid
